+++ frontend/components/background/StarBg.tsx
